examples: Correct allocation of StatResult
The type StatResult defined as a pointer to a stat_obj_struct structure
(sStatObject is also defined as this structure). When memory is
allocated for r in main(), the size of StatResult is erroneously used
instead of the size of sStatResult. This results in a heap buffer
overflow when stat_obj_value() is called, as it expects r to be a
pointer to a stat_obj_struct structure.